CIV-11185 moving hearingFee own directory
